Blur, sharpen and enhance filters are ok

+++ Filters_SDL/src/Matrix.c
@@ -38,7 +38,7 @@ void PrintMatrix(struct Matrix *m)
 	{
 		for (int j = 0; j < m -> cols; j++)
 		{
-			printf("%i ",(int) *(m -> mat + i * m -> cols + j));
+			printf("%3f ", *(m -> mat + i * m -> cols + j));
 		}
 	printf("\n");
 	}
